+++ src/main/scala/com/sfxcode/sapphire/core/control/FXTableValueFactory.scala
@@ -5,8 +5,7 @@ import javafx.scene.control.TableColumn
 import javafx.scene.control.TableColumn.CellDataFeatures
 import javafx.util.Callback
 
-@deprecated(since = "1.7.4", message = "Use FXTableValueFactory")
-class FXValueFactory[S <: AnyRef, T]
+class FXTableValueFactory[S <: AnyRef, T]
   extends Callback[TableColumn.CellDataFeatures[S, T], ObservableValue[T]]
   with TableValue {